+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `--dump-config` CLI flag that dumps the Misti configuration file in use: PR [#79](https://github.com/nowarp/misti/pull/79)
 - Naming convention to skip unused identifiers: PR [#82](https://github.com/nowarp/misti/pull/82)
+- `--all-detectors` CLI flag activates all the available built-in detectors, regardless of whether they are selected in the config
 
 ### Changed
 - IRBuilder: Mark nodes without successors as `Exit` kind: PR [#80](https://github.com/nowarp/misti/pull/80)